// ClientApp/src/app/Models/paciente.model.ts
export class Paciente
{
    id: number = 0;    
    nome: string = "";
    sobrenome: string = "";
    prontuario: string = "";
    dataNascimento: Date = new Date();
    genero: string = "";
    cpf: string = "";
    rg: string = "";
    ufRG: string = "";
    email: string = "";
    celular: string = "";
    telefone: string = "";
    convenio: string = "";
    carteira: string = "";
    carteiraValidade: Date = new Date();
}